Split out windows and non-windows
